DATA: Align CSV IDs with DB schema across services

+++ services/booking/db_setup.sql
@@ -3,8 +3,8 @@ DROP TABLE IF EXISTS restaurant_slots;
 
 CREATE TABLE IF NOT EXISTS booking(
     id serial PRIMARY KEY,
-    user_id INT NOT NULL,
-    restaurant_id INT NOT NULL,
+    user_id NUMERIC(22, 0) NOT NULL,
+    restaurant_id NUMERIC(22, 0) NOT NULL,
     time_start TIMESTAMP NOT NULL,
     time_end TIMESTAMP NOT NULL,
     people_count INT NOT NULL
@@ -12,7 +12,7 @@ CREATE TABLE IF NOT EXISTS booking(
 
 CREATE TABLE IF NOT EXISTS restaurant_slots(
     id SERIAL PRIMARY KEY,
-    restaurant_id INT NOT NULL,
+    restaurant_id NUMERIC(22, 0) NOT NULL,
     time_start TIMESTAMP NOT NULL,
     time_end TIMESTAMP NOT NULL,
     max_slots INT NOT NULL,
