+++ services/sponsor/db_setup.sql
@@ -1,11 +1,11 @@
 CREATE TABLE IF NOT EXISTS sponsorship(
-    restaurant_id INT PRIMARY KEY,
+    restaurant_id NUMERIC(22, 0) PRIMARY KEY,
     tier INT,
     until DATE
 );
 
 CREATE TABLE IF NOT EXISTS restaurant_categories (
     id SERIAL PRIMARY KEY,
-    restaurant_id INT NOT NULL REFERENCES sponsorship(restaurant_id) ON DELETE CASCADE,
+    restaurant_id NUMERIC(22, 0) NOT NULL REFERENCES sponsorship(restaurant_id) ON DELETE CASCADE,
     category TEXT NOT NULL
 );
